chore(cli): remove integration command, consolidate into git command
- Remove integration command (integration github connect/disconnect)
- Move integration/api.ts to git/api.ts
- Update all imports to use the new location
- The git command already has full feature parity:
  - git account add = integration github connect
  - git account remove = integration github disconnect
