+++ packages/cli/src/cmd/git/unlink.ts
@@ -2,7 +2,7 @@ import { createSubcommand } from '../../types';
 import * as tui from '../../tui';
 import { getCommand } from '../../command-prefix';
 import { ErrorCode } from '../../errors';
-import { getProjectGithubStatus, unlinkProjectFromRepo } from '../integration/api';
+import { getProjectGithubStatus, unlinkProjectFromRepo } from './api';
 import { z } from 'zod';
 
 const UnlinkOptionsSchema = z.object({
