+++ packages/cli/src/cmd/index.ts
@@ -12,7 +12,6 @@ export async function discoverCommands(): Promise<CommandDefinition[]> {
 		import('./dev').then((m) => m.command),
 		import('./git').then((m) => m.gitCommand),
 		import('./help').then((m) => m.command),
-		import('./integration').then((m) => m.command),
 		import('./profile').then((m) => m.command),
 		import('./project').then((m) => m.command),
 		import('./repl').then((m) => m.command),
